requirements.txt
numpy
pillow
requests
